summarization-service/requirements.txt
flask
langchain
langchain-openai  # Changed from langchain-community
python-dotenv